Correct potential rounding error

+++ lib/tables/glyf.c
@@ -753,14 +753,14 @@ static void glyf_write_simple(glyf_glyph *g, caryll_buffer *gbuf) {
 	bufclear(flags);
 	bufclear(xs);
 	bufclear(ys);
-	float cx = 0;
-	float cy = 0;
+	int16_t cx = 0;
+	int16_t cy = 0;
 	for (uint16_t cj = 0; cj < g->numberOfContours; cj++) {
 		for (uint16_t k = 0; k < g->contours[cj].pointsCount; k++) {
 			glyf_point *p = &(g->contours[cj].points[k]);
 			uint8_t flag = (p->onCurve & MASK_ON_CURVE) ? GLYF_FLAG_ON_CURVE : 0;
-			int16_t dx = p->x - cx;
-			int16_t dy = p->y - cy;
+			int16_t dx = (int16_t)(p->x) - cx;
+			int16_t dy = (int16_t)(p->y) - cy;
 			if (dx == 0) {
 				flag |= GLYF_FLAG_SAME_X;
 			} else if (dx >= -0xFF && dx <= 0xFF) {
